🚚 api: move files to src directory
