+++ api/lib/application/complementary-certifications/index.js
@@ -30,29 +30,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/admin/complementary-certifications/{id}/target-profiles',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.adminMemberHasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: complementaryCertificationController.getComplementaryCertificationTargetProfileHistory,
-        tags: ['api', 'admin'],
-        notes: [
-          'Cette route est restreinte aux utilisateurs authentifiés avec le rôle Super Admin, Support et Métier',
-          'Elle renvoie les informations du profil cible courant de la certification complémentaire.',
-        ],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/admin/complementary-certifications/attachable-target-profiles',
       config: {
         validate: {
