+++ api/lib/domain/usecases/index.js
@@ -80,7 +80,6 @@ import * as competenceTreeRepository from '../../infrastructure/repositories/com
 import * as complementaryCertificationCourseResultRepository from '../../infrastructure/repositories/complementary-certification-course-result-repository.js';
 import * as complementaryCertificationHabilitationRepository from '../../infrastructure/repositories/complementary-certification-habilitation-repository.js';
 import * as complementaryCertificationRepository from '../../infrastructure/repositories/complementary-certification-repository.js';
-import * as complementaryCertificationTargetProfileHistoryRepository from '../../infrastructure/repositories/complementary-certification-target-profile-history-repository.js';
 import * as countryRepository from '../../infrastructure/repositories/country-repository.js';
 import * as courseRepository from '../../infrastructure/repositories/course-repository.js';
 import * as cpfCertificationResultRepository from '../../infrastructure/repositories/cpf-certification-result-repository.js';
@@ -290,7 +289,6 @@ const dependencies = {
   complementaryCertificationCourseResultRepository,
   complementaryCertificationHabilitationRepository,
   complementaryCertificationRepository,
-  complementaryCertificationTargetProfileHistoryRepository,
   config,
   correctionRepository: repositories.correctionRepository,
   countryRepository,
