+++ api/src/certification/complementary-certification/application/complementary-certification-controller.js
@@ -1,0 +1,15 @@
+import { usecases } from '../../shared/domain/usecases/index.js';
+import * as complementaryCertificationSerializer from '../infrastructure/serializers/jsonapi/complementary-certification-serializer.js';
+
+const getComplementaryCertificationTargetProfileHistory = async function (request) {
+  const complementaryCertificationId = request.params.id;
+  const complementaryCertification = await usecases.getComplementaryCertificationTargetProfileHistory({
+    complementaryCertificationId,
+  });
+  return complementaryCertificationSerializer.serializeForAdmin(complementaryCertification);
+};
+
+const complementaryCertificationController = {
+  getComplementaryCertificationTargetProfileHistory,
+};
+export { complementaryCertificationController };
