+++ api/src/certification/complementary-certification/domain/usecases/get-complementary-certification-target-profile-history.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertificationTargetProfileHistory } from '../../domain/models/ComplementaryCertificationTargetProfileHistory.js';
+import { ComplementaryCertificationTargetProfileHistory } from '../models/ComplementaryCertificationTargetProfileHistory.js';
 
 const getComplementaryCertificationTargetProfileHistory = async function ({
   complementaryCertificationId,
