+++ api/src/certification/complementary-certification/infrastructure/serializers/jsonapi/complementary-certification-serializer.js
@@ -1,0 +1,17 @@
+import jsonapiSerializer from 'jsonapi-serializer';
+
+const { Serializer } = jsonapiSerializer;
+
+const serializeForAdmin = function (complementaryCertification) {
+  return new Serializer('complementary-certification', {
+    attributes: ['label', 'hasExternalJury', 'targetProfilesHistory'],
+    targetProfilesHistory: {
+      attributes: ['id', 'name', 'attachedAt', 'detachedAt', 'badges'],
+      badges: {
+        attributes: ['id', 'label', 'level'],
+      },
+    },
+  }).serialize(complementaryCertification);
+};
+
+export { serializeForAdmin };
