+++ api/src/certification/complementary-certification/routes.js
@@ -1,5 +1,7 @@
 import * as attachTargetProfile from './application/attach-target-profile-route.js';
+import * as complementaryCertification from './application/complementary-certification-route.js';
 
 const attachTargetProfileRoutes = [attachTargetProfile];
+const complementaryCertificationRoutes = [complementaryCertification];
 
-export { attachTargetProfileRoutes };
+export { attachTargetProfileRoutes, complementaryCertificationRoutes };
