+++ api/tests/tooling/domain-builder/factory/build-complementary-certification-target-profile-history-for-admin.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertificationTargetProfileHistory } from '../../../../lib/domain/models/ComplementaryCertificationTargetProfileHistory.js';
+import { ComplementaryCertificationTargetProfileHistory } from '../../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationTargetProfileHistory.js';
 
 const buildComplementaryCertificationTargetProfileHistory = function ({
   id = 1,
